cosas

+++ src/Library/Interfaces/IMedia.cs
@@ -3,7 +3,7 @@ namespace Library
     public interface IMedia
     {
         string Name {get;}
-        float Calification {get;}
+        double Calification {get;}
         List<Tag> Tags {get;}
     }
 }
